+++ components/atoms/index.ts
@@ -16,5 +16,5 @@ export * from "./NumberField/NumberField"
 export * from './Tooltip/Tooltip';
 export * from './LoadingButton/LoadingButton';
 export * from './Checkbox/Checkbox';
-export * from './Toggleswitch/Toggleswitch';
+export * from './Switch/Switch';
 export * from './Dialog/Dialog';
